gateway: add subscription gateway endpoint utils #260
